+++ webui/react/src/pages/ExperimentDetails/ExperimentVisualization.tsx
@@ -22,7 +22,7 @@ import {
   HpImportanceMap, HpImportanceMetricMap, HyperparameterType, MetricName, MetricType, RunState,
   Scale,
 } from 'types';
-import { hpImportanceSorter } from 'utils/sort';
+import { hpImportanceSorter } from '../../utils/experiment';
 
 import css from './ExperimentVisualization.module.scss';
 import ExperimentVisualizationFilters, {
