+++ webui/react/src/utils/experiment.ts
@@ -1,27 +1,27 @@
 import {
-  cancellableRunStates,
-  deletableRunStates,
-  killableRunStates,
-  pausableRunStates,
-  terminalRunStates,
+    cancellableRunStates,
+    deletableRunStates,
+    killableRunStates,
+    pausableRunStates,
+    terminalRunStates,
 } from 'constants/states';
 import { RawJson } from 'shared/types';
-import { clone, deletePathList, getPathList, isNumber, setPathList,
-  unflattenObject } from 'shared/utils/data';
+import { clone, deletePathList, getPathList, isNumber, setPathList, unflattenObject } from 'shared/utils/data';
 import {
-  AnyTask,
-  DetailedUser,
-  ExperimentAction,
-  ExperimentBase,
-  ExperimentItem,
-  ExperimentSearcherName,
-  Hyperparameters,
-  HyperparameterType,
-  Project,
-  ProjectExperiment,
-  RunState,
-  TrialDetails,
-  TrialHyperparameters,
+    AnyTask,
+    DetailedUser,
+    ExperimentAction,
+    ExperimentBase,
+    ExperimentItem,
+    ExperimentSearcherName,
+    HpImportance,
+    Hyperparameters,
+    HyperparameterType,
+    Project,
+    ProjectExperiment,
+    RunState,
+    TrialDetails,
+    TrialHyperparameters,
 } from 'types';
 
 type ExperimentChecker = (
@@ -217,3 +217,27 @@ export const getProjectExperimentForExperimentItem = (
     workspaceId: project?.workspaceId ?? 0,
     workspaceName: project?.workspaceName,
   } as ProjectExperiment);
+export const runStateSortValues: Record<RunState, number> = {
+    [RunState.Active]: 0,
+    [RunState.Paused]: 1,
+    [RunState.StoppingError]: 2,
+    [RunState.Errored]: 3,
+    [RunState.StoppingCompleted]: 4,
+    [RunState.Completed]: 5,
+    [RunState.StoppingCanceled]: 6,
+    [RunState.Canceled]: 7,
+    [RunState.Deleted]: 7,
+    [RunState.Deleting]: 7,
+    [RunState.DeleteFailed]: 7,
+    [RunState.Unspecified]: 8,
+};
+export const hpImportanceSorter = (a: string, b: string, hpImportance: HpImportance): number => {
+    const aValue = hpImportance[a];
+    const bValue = hpImportance[b];
+    if (aValue < bValue) return 1;
+    if (aValue > bValue) return -1;
+    return 0;
+};
+export const runStateSorter = (a: RunState, b: RunState): number => {
+    return runStateSortValues[a] - runStateSortValues[b];
+};
